Fix unexpected lead number missing (#10)
* Fix unexpect lead number missing

In pervious code, if `i = 0`, ++ternary[i - 1] couldn't take affect.  
So my solution is manully insert a zero element, though it's slow - O(n).
You can choose to reverse the whole vector at first, and then the addition will be happened on right side, which means you can use `push_back()`.

* Change the direction of vector

I redirected the way digits stored in vector.

In the usual implementation, ternary numbers are represented using vectors, with each int representing one ternary bit of the number. 

When data being read into a vector, the highest bit of the number is at the head of the vector (the position with the smallest subscript). The reason for this is that **the length of ternary numbers may change** , but we want the same weight bits to remain aligned (e.g., we want all the digits to be in subscript [0], all the tens to be in subscript [1]...); at the same time, we want all the tens to be in subscript [1]... ...); and the fact that addition, subtraction, and multiplication are generally performed from the first digit (think of vertical arithmetic in elementary school) gives a good reason for "inverted storage". And this makes benifits for further numeric computations in this project.

* clang-format edit

When `DerivePointerAlignment` enabled. `PointerAlignment` is then used only as fallback, which breaks the style.

* have brackets, keeps things easy to read.

Co-authored-by: Brandon P. <pacewicbrandon@gmail.com>

---------

Co-authored-by: Brandon P. <pacewicbrandon@gmail.com>